+++ src/config/routerPage.ts
@@ -4,4 +4,5 @@ export const RouterPage: Record<RouterKey.RouterName, RouteComponent> = {
   workbench: () => import("@/views/home/DashBoard.vue"),
   chatGpt: () => import("@/views/gpt/ChatGpt.vue"),
   chatStream: () => import("@/views/gpt/ChatStream.vue"),
+  userInfo: () => import("@/views/auth/UsersInfo.vue"),
 };