+++ src/layouts/default/View.vue
@@ -2,7 +2,7 @@
 import NavigationDrawer from "../main/NavigationDrawer.vue";
 import AppBar from "../main/AppBar.vue";
 import AppMain from "../main/AppMain.vue";
-import SnackBar from "@/components/SnackBar.vue";
+import SnackBar from "@/components/common/SnackBar.vue";
 </script>
 <template>
   <v-layout>