+++ src/store/open/openApp.ts
@@ -4,6 +4,7 @@ import {
   OpenState,
   Open_KEY,
   CLEAN_APP_SETTING,
+  SAVE_APP_STATE,
 } from "./help";
 
 export const openAppStore = defineStore("openApp", {