+++ src/views/gpt/components/MarkdownMsg.vue
@@ -1,0 +1,14 @@
+<script setup lang="ts">
+import "highlight.js/styles/base16/ir-black.css";
+import MarkdownIt from "markdown-it";
+import Markdown from "vue3-markdown-it";
+defineProps({
+  content: {
+    type: String,
+    default: "",
+  },
+});
+</script>
+<template>
+  <Markdown :source="content" :breaks="true" />
+</template>