+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## Unreleased (2023-04-27)
+## Unreleased (2023-05-04)
 
+* ✨ feat(custom): 添加apiModel自定义 ([12c882a](https://github.com/saofeng-cyber/vuetify-project/commit/12c882a))